Force C++17
Node 20.18.0 seems to have changed headers that's causing node-gyp to not detect this should be compiled with C++11 support. For now, just force C++17 so we don't run into this in the future if this isn't fixed by the time Node starts to require C++14 or C++17. The future is now.